+++ pkg/provider/application_source_from_template_resource_test.go
@@ -45,8 +45,8 @@ func testApplicationSourceFromTemplateResource(t *testing.T, isMock bool) {
 
 	newName := "new_from_template " + nameSalt
 
-	baseEnvironmentID := testStreamlitBaseEnvID
-	baseEnvironmentVersionID := testAppSourceBaseEnvVersionID
+	baseEnvironmentID := testAppBaseEnvID
+	baseEnvironmentVersionID := testAppSourceBaseEnvVersionID(t)
 
 	appCodeFileName := "flask_app.py"
 	appCode := `import streamlit as st